fix warnings

+++ dnvim/d3d/constantbuffer.cpp
@@ -47,7 +47,7 @@ public:
     void AddCBSlot(const Microsoft::WRL::ComPtr<ID3D11Device> &device, size_t size)
     {
 		D3D11_BUFFER_DESC desc = { 0 };
-		desc.ByteWidth = size;
+		desc.ByteWidth = (UINT)size;
 		desc.Usage = D3D11_USAGE_DEFAULT;
 		desc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
 		Microsoft::WRL::ComPtr<ID3D11Buffer> pBuffer;
@@ -107,13 +107,13 @@ bool ConstantBuffer::Initialize(const Microsoft::WRL::ComPtr<ID3D11Device> &pDev
     pReflector->GetDesc(&shaderdesc);
 
     // analize constant buffer
-    for (size_t i = 0; i < shaderdesc.ConstantBuffers; ++i){
+    for (UINT i = 0; i < shaderdesc.ConstantBuffers; ++i){
         auto cb = pReflector->GetConstantBufferByIndex(i);
         D3D11_SHADER_BUFFER_DESC desc;
         cb->GetDesc(&desc);
 		LOGD << "[" << i << ": " << desc.Name << "]";
         impl->AddCBSlot(pDevice, desc.Size);
-        for (size_t j = 0; j < desc.Variables; ++j){
+        for (UINT j = 0; j < desc.Variables; ++j){
             auto v = cb->GetVariableByIndex(j);
             D3D11_SHADER_VARIABLE_DESC vdesc;
             v->GetDesc(&vdesc);
@@ -122,7 +122,7 @@ bool ConstantBuffer::Initialize(const Microsoft::WRL::ComPtr<ID3D11Device> &pDev
         }
     }
 
-	for (size_t i = 0; i < shaderdesc.BoundResources; ++i){
+	for (UINT i = 0; i < shaderdesc.BoundResources; ++i){
 		D3D11_SHADER_INPUT_BIND_DESC desc;
 		pReflector->GetResourceBindingDesc(i, &desc);
 		switch (desc.Type)
